tratamento de excecoes - delete

+++ src/main/java/com/projeto/ecommerce/services/UserService.java
@@ -2,8 +2,10 @@ package com.projeto.ecommerce.services;
 
 import com.projeto.ecommerce.entities.User;
 import com.projeto.ecommerce.repositories.UserRepository;
+import com.projeto.ecommerce.services.exceptions.DatabaseException;
 import com.projeto.ecommerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,8 +30,15 @@ public class UserService {
         return repository.save(obj);
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try{
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
